+++ lib/model-file.ts
@@ -11,7 +11,7 @@ const JSON_EXT: string = '.json';
 export const SUPPORT_EXT: string[] = [JSON_EXT, YAML_EXT];
 
 export class ModelFile {
-  public static Read(file: string): def.ModelInfo {
+  public static read(file: string):any {
     let ext = path.extname(file);
     if (JSON_EXT === ext) {
       return $.readJSONWithComments(file);
@@ -22,4 +22,11 @@ export class ModelFile {
       throw Error(`Not support '${ext}'`);
     }
   }
+
+  public static findFile(fileWithoutExt: string): string | boolean   {
+    if (fs.existsSync(fileWithoutExt + JSON_EXT)) return fileWithoutExt + JSON_EXT;
+    else if (fs.existsSync(fileWithoutExt + YAML_EXT)) return fileWithoutExt + YAML_EXT;
+    else return false;
+  }
+
 }