+++ lib/regexp-common.ts
@@ -1,0 +1,4 @@
+export namespace RegExpCommon {
+  export const NAME_SPACE: RegExp = /[\w-/_]+/;
+  export const NAME_SPACE_ALLOW: string = "'-', '_' , '/'";
+}