+++ test/items/mcgen-config/test1/__check__/ts/model.ts
@@ -1,0 +1,21 @@
+import { Promise } from "./System";
+import { ExType } from "./Sys/Data";
+
+
+
+export interface Mandor {
+  /**
+   * Required. 
+   */
+  prop0: string;
+  /**
+   * Required. 
+   */
+  prop1: Promise;
+  /**
+   * Required. 
+   */
+  prop2: ExType;
+}
+
+